feat(calendar): refactor task imports to streamline component usage across day, month, and week views
