+++ src/components/calendar/index.ts
@@ -2,6 +2,3 @@ export { CalendarHeader } from "./calendar-header";
 export { MonthView } from "./month-view";
 export { DayView } from "./day-view";
 export { WeekView } from "./week-view";
-export { TaskDetailsDialog } from "./task-details-dialog";
-export { TaskDialog } from "./task-dialog";
-export { AddTaskButton } from "./add-task-button";
