+++ src/components/calendar/month-view.tsx
@@ -3,9 +3,8 @@ import { useState } from "react";
 import { useCalendar } from "@/context/calendar-context";
 import { generateMonthCalendar } from "@/lib/calendar-helpers";
 import { cn } from "@/lib/utils";
-import { TaskDetailsDialog } from "./task-details-dialog";
+import { TaskDetailsDialog, DraggableTask } from "@/components/task";
 import { DroppableDayCell } from "./dnd/droppable-day-cell";
-import { DraggableTask } from "./dnd/draggable-task";
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog";
 import { Badge } from "@/components/ui/badge";
 import type { CalendarTask, TaskColor } from "@/types/calendar";
